unix: remove StTimespec type on AIX
On AIX, Stat_t's fields dealing with time are of type StTimespec while
all other GOOS are using Timespec.
StTimespec and Timespec are the same on ppc but not in ppc64. Therefore,
values returned by ppc64 syscalls need to be adjusted in order to
allow the use of Timespec instead of StTimespec.

Fixes golang/go#32073

Reviewed-on: https://go-review.googlesource.com/c/sys/+/177838

+++ unix/syscall_aix_test.go
@@ -107,10 +107,10 @@ func TestUtimesNanoAt(t *testing.T) {
 		t.Fatalf("Lstat: %v", err)
 	}
 	if runtime.GOARCH == "ppc64" {
-		if int64(st.Atim.Sec) != int64(ts[0].Sec) || st.Atim.Nsec != int32(ts[0].Nsec) {
+		if int64(st.Atim.Sec) != int64(ts[0].Sec) || st.Atim.Nsec != ts[0].Nsec {
 			t.Errorf("UtimesNanoAt: wrong atime: %v", st.Atim)
 		}
-		if int64(st.Mtim.Sec) != int64(ts[1].Sec) || st.Mtim.Nsec != int32(ts[1].Nsec) {
+		if int64(st.Mtim.Sec) != int64(ts[1].Sec) || st.Mtim.Nsec != ts[1].Nsec {
 			t.Errorf("UtimesNanoAt: wrong mtime: %v", st.Mtim)
 		}
 	} else {
